Move static functions to Utils
